+++ app/src/main/java/woowacourse/shopping/view/productdetail/ProductDetailViewModel.kt
@@ -7,9 +7,9 @@ import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.ViewModelProvider.AndroidViewModelFactory.Companion.APPLICATION_KEY
 import androidx.lifecycle.createSavedStateHandle
 import androidx.lifecycle.viewmodel.CreationExtras
+import woowacourse.shopping.data.Cart
+import woowacourse.shopping.data.CartImpl
 import woowacourse.shopping.model.products.Product
-import woowacourse.shopping.view.cart.Cart
-import woowacourse.shopping.view.cart.CartImpl
 
 class ProductDetailViewModel(
     private val cart: Cart,
